% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANNt_OCA2023.R
\name{ANNt_OCA2023}
\alias{ANNt_OCA2023}
\title{ANNt OCA 2023
reproduce Oliveira, Ceretta, & Albrecht (2023)}
\usage{
ANNt_OCA2023(
  Tickers,
  RM,
  Rf,
  Initial_Date,
  Final_Date_Training,
  Final_Date,
  Periodicity,
  Hidden,
  Stepmax,
  N_Assets
)
}
\arguments{
\item{Tickers}{Current SP500 Tickers on 2022-07-12}

\item{RM}{Proxy of the market: SP500}

\item{Initial_Date}{Series start Date: 2018-01-10}

\item{Final_Date_Training}{Training series start Date: 2020-01-20}

\item{Final_Date}{End date of the treatment series: 2022-07-12}

\item{Periodicity}{should be one of “daily”, “weekly”, “monthly”, “hourly”, “1minutes”, “2minutes”, “5minutes”, “15minutes”, “30minutes”, “60minutes”, “90minutes”. (Intraday maximum 7 days)}

\item{Hidden}{Number of hidden neurons (If ” is the length series). For a good performance use '' to form a square input x hidden matrix of neurons}

\item{Stepmax}{Number of replications per asset to train the ANN. For a good performance, use 7500}

\item{N_Assets}{Limit of asset numbers in the portfolio}
}
\description{
For more details see:
Oliveira, A. S., Ceretta, P. S., & Albrecht, P. (2023). Performance comparison of multifractal techniques and artificial neural networks in the construction of investment portfolios. Finance Research Letters, 55, 103814.
}
\examples{
Tickers <-'Current_SP500_Tickers'
RM <-c('^GSPC') #RM the S&P500
Rf <- 0.0311
Initial_Date <-c('2018-01-10')
Final_Date_Training <- c('2020-01-20')
Final_Date <-c('2022-07-12')
Periodicity <- c('daily')
Hidden <- ''
Stepmax <- 7500
N_Assets <- 5
ANNt:::ANNt_OCA2023('Current_SP500_Tickers', '^GSPC', 0.0311, '2018-01-10', '2020-01-20', '', 'daily','',7500,5)
}
