% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backup_ANNt.R
\name{Backup_ANNt}
\alias{Backup_ANNt}
\title{Backup_ANNt
Generate backup of the role}
\usage{
Backup_ANNt()
}
\arguments{
\item{()}{No require parameters}
}
\description{
Backup_ANNt
Generate backup of the role
}
\examples{
Backup_ANNt()
}
