% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_OCA2023.R
\name{Data_OCA2023}
\alias{Data_OCA2023}
\title{OCA 2023
Generate the results of Oliveira, Ceretta & Albrecht (2023)}
\usage{
Data_OCA2023()
}
\arguments{
\item{()}{No require parameters}
}
\description{
For more details see: Oliveira, A. S., Ceretta, P. S., & Albrecht, P. (2023). Performance comparison of multifractal techniques and artificial neural networks in the construction of investment portfolios. Finance Research Letters, 55, 103814.
}
\examples{
Data_OCA2023()

}
\references{
Oliveira, A. S., Ceretta, P. S., & Albrecht, P. (2023). Performance comparison of multifractal techniques and artificial neural networks in the construction of investment portfolios. Finance Research Letters, 55, 103814.
}
\author{
Alexandre Silva de Oliveira
}
