% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen_efficient_frontier.R
\name{Gen_efficient_frontier}
\alias{Gen_efficient_frontier}
\title{Gen_efficient_frontier
Generate efficient frontier of Markowitz}
\usage{
Gen_efficient_frontier(Initial_Analysis_Date, Final_Analysis_Date)
}
\arguments{
\item{Initial_Analysis_Date}{Initial Date of Analysis. If '' is the same of ANNt Testing}

\item{Final_Analysis_date}{Final Date of Analysis. If '' is the same of ANNt testing}
}
\description{
Gen_efficient_frontier
Generate efficient frontier of Markowitz
}
\examples{
Initial_Analysis_Date <- c('2023-01-03')
Final_Analysis_Date <- c('2023-01-03')
Gen_efficient_frontier('','')
}
