% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Out_of_sample.R
\name{Out_of_sample}
\alias{Out_of_sample}
\title{Out of Sample}
\usage{
Out_of_sample(Initial_Date_Out, Final_Date_Out)
}
\arguments{
\item{Initial_Date_Out}{Initial Date Out-of-sample}

\item{Final_Date_Out}{Final Date Out-of-sample}
}
\description{
Performs an out-of-sample run with the weights defined within the sample.
You need to generate portfolios first with the Gen_portfolio() command.
}
\examples{
Initial_Date_Out <- c('2023-08-03')
Final_Date_Out <- c('')
# Generate performs out-of-sample
Out_of_sample('2023-08-03','')

}
