% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_CUSUM.R
\name{Plot_CUSUM}
\alias{Plot_CUSUM}
\title{Plot_CUSUM
Prot graphic Cumulative Sum of Returns}
\usage{
Plot_CUSUM(Change, Control)
}
\arguments{
\item{Change}{Magnitude of change in the cumulative sum measured as the accepted standard deviation (sd) amount. Default is 1 sd}

\item{Control}{Control limit as a multiple of the sd. Default is 5 sd}
}
\description{
Plot_CUSUM
Prot graphic Cumulative Sum of Returns
}
\examples{
Plot_CUSUM(1,5)
}
