% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_New_efficient_frontier.R
\name{Plot_New_efficient_frontier}
\alias{Plot_New_efficient_frontier}
\title{Plot_New_efficient_frontier
Generate the new efficient frontier}
\usage{
Plot_New_efficient_frontier()
}
\arguments{
\item{()}{No require parameters}
}
\description{
Plot_New_efficient_frontier
Generate the new efficient frontier
}
\examples{
Plot_New_efficient_frontier()
}
