% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio_backtest.R
\name{Portfolio_backtest}
\alias{Portfolio_backtest}
\title{Portfolio_backtest
Generates the portfolio backtest}
\usage{
Portfolio_backtest()
}
\arguments{
\item{()}{No required parameter}
}
\description{
Portfolio_backtest
Generates the portfolio backtest
}
\examples{
Portfolio_backtest()


}
