% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio_backtesting.R
\name{Portfolio_backtesting}
\alias{Portfolio_backtesting}
\title{Portfolio_backtesting
Realize the portfolio backtest.}
\usage{
Portfolio_backtesting()
}
\arguments{
\item{()}{no required parameters}
}
\description{
Portfolio_backtesting
Realize the portfolio backtest.
}
\examples{
Portfolio_backtesting()

}
