% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Specify_Pf_RM.R
\name{Specify_Pf_RM}
\alias{Specify_Pf_RM}
\title{Specify RM
permit that create a specific portfolio benchmark to use in ANNt}
\usage{
Specify_Pf_RM(Name, Portfolio, Weights)
}
\arguments{
\item{Name}{Name of the portfolio}

\item{Portfolio}{Describes the assets in the portfolio}

\item{Weights}{Describe the weights of the portfolio´s assets}
}
\description{
Portfolio generate with the assets and weights that you specify
}
\examples{
Name = c("My_Pf")
Portfolio=c("AAPL", "BAC", "CVX", "KO", "AXP", "KHC", "OXY", "MCO") # PORTFOLIO´s Buffet 2023
Weights=c( 0.441, 0.089, 0.079, 0.075, 0.074, 0.038, 0.038, 0.023) # PORTFOLIO weights 2023
Specify_Pf_RM(Name,Portfolio,Weights)
}
