% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_New_efficient_frontier.R
\name{Plot_New_efficient_frontier}
\alias{Plot_New_efficient_frontier}
\title{Plot New efficient frontier:
generate the new efficient frontier}
\usage{
Plot_New_efficient_frontier()
}
\arguments{
\item{()}{No require parameters}
}
\description{
This chart presents the assets most likely to outperform the benchmark.
The selection of assets must occur from right to left, as indicated by the arrow.
Allows you to select the assets that should be included in the portfolio to obtain high performance.
}
\examples{
Plot_New_efficient_frontier()
}
