% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio_backtesting.R
\name{Portfolio_backtesting}
\alias{Portfolio_backtesting}
\title{Portfolio_backtesting
Realize the portfolio backtest.}
\usage{
Portfolio_backtesting(Date_Initial_Backtesting, Date_Final_Backtesting)
}
\arguments{
\item{Date_Initial_Backtesting}{Date initial of the backtest}

\item{Date_Final_Backtesting}{Date final of the backtest}
}
\description{
Portfolio_backtesting
Realize the portfolio backtest.
}
\examples{
Date_Initial_Backtesting =''
Date_Final_Backtesting =''
Portfolio_backtesting('', '')

}
